codegen.c: genop_3 to take `uint16_t` instead of `uint8_t`.

+++ mrbgems/mruby-compiler/core/codegen.c
@@ -287,10 +287,10 @@ genop_2(codegen_scope *s, mrb_code i, uint16_t a, uint16_t b)
 }
 
 static void
-genop_3(codegen_scope *s, mrb_code i, uint16_t a, uint16_t b, uint8_t c)
+genop_3(codegen_scope *s, mrb_code i, uint16_t a, uint16_t b, uint16_t c)
 {
   genop_2(s, i, a, b);
-  gen_B(s, c);
+  gen_B(s, (uint8_t)c);
 }
 
 static void
